# application/services/custom_services/newtest/src/models/entitymodels/Newon.ts
export interface Newon 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   Nmae: { type: String },
   EmailId: { type: String }
}